Add support for the TC_AWS_ENDPOINT environment variable
Allows using Thumbor with other S3-compatible services, such as:
- Minio
- Google Cloud Platform's storage system

The TC_AWS_ENDPOINT value defaults to None.

If some other (string) value is to be provided, it's expected that
the actual value is wrapped in quotes. This is consistent with
the handling of other such "quote-wrapped-string or None" options, like:
- ERROR_FILE_LOGGER
- HTTP_LOADER_PROXY_HOST
- etc. (many others)

The ternary operator if-statement is there, so that
an empty string ('') in the TC_AWS_ENDPOINT environment variable
would also yield `TC_AWS_ENDPOINT = None`.