Add HTTP_LOADER_MAX_CLIENTS to README.md
I've noticed that the README was missing the `HTTP_LOADER_MAX_CLIENTS` variable. I've tested it on a 6.3.0 image, and by default the value is 10:

```
$ docker run -it apsl/thumbor:6.3.0 cat thumbor.conf | grep HTTP_LOADER_MAX_CLIENTS
HTTP_LOADER_MAX_CLIENTS = 10
```
Passing it as an environment variable actually sets it as expected:


```
$ docker run -it -e HTTP_LOADER_MAX_CLIENTS=99 apsl/thumbor:6.3.0 cat thumbor.conf | grep HTTP_LOADER_MAX_CLIENTS
HTTP_LOADER_MAX_CLIENTS = 99
```

I assumed you just forgot about it, but if there's any reason why this wasn't originally included in the README please let me know! Since it was such a simple change I though it'd be best to open a PR instead of asking.